# perf(blooms): Replace JSON lib for encoding/decoding metas